third_party/WebKit/LayoutTests/bluetooth/characteristic/getDescriptors/blocklisted-descriptors-not-present.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Making sure no blocklisted descriptors are present.';

bluetooth_test(() => setUpBlocklistDescriptor()
    .then(({characteristic}) => characteristic.getDescriptors())
    .then(descriptors => descriptors.forEach(descriptor =>
        // bad_uuid is a blocklisted UUID.  It should never be present
        // in a getDescriptors() result.
        assert_true(descriptor.uuid !== blocklist_test_descriptor_uuid))),
    test_desc);
</script>
